chore: Added rollup bundling to client side javascript
