+++ apps/design-system/_scripts/main.ts
@@ -13,15 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-const themeSwitch = document.querySelector('.ds-theme-switch');
-if (themeSwitch) {
-  themeSwitch.addEventListener('click', () => {
-    if (document.body.classList.contains('fluid-theme--abyss')) {
-      document.body.classList.remove('fluid-theme--abyss');
-      document.body.classList.add('fluid-theme--surface');
-    } else {
-      document.body.classList.add('fluid-theme--abyss');
-      document.body.classList.remove('fluid-theme--surface');
-    }
-  });
-}
+
+import { themeSwitcher } from './src/theme-switch';
+themeSwitcher();
