+++ package.json
@@ -103,6 +103,7 @@
     "@angular/router": "^10.0.7",
     "@bazel/bazelisk": "^1.5.0",
     "@bazel/ibazel": "^0.13.1",
+    "@bazel/rollup": "^2.0.1",
     "@bazel/typescript": "^2.0.0",
     "@commitlint/cli": "^8.2.0",
     "@nguniversal/builders": "^10.0.1",
